requirements.txt
discord.py
PyNaCl
spotipy